fix: 오타수정, 패키지 변경, Setter제거
